Move all non-accordion section code out of AccordionSection
