+++ src/scenes/Products/ProductForm/ProductForm.tsx
@@ -1,13 +1,39 @@
+import { makeStyles, Typography } from '@material-ui/core';
+import { useSnackbar } from 'notistack';
 import React from 'react';
 import { Form, FormProps } from 'react-final-form';
+import { useNavigate } from 'react-router';
+import { GQLOperations } from '../../../api';
+import { ErrorButton } from '../../../components/ErrorButton';
+import {
+  FieldGroup,
+  SubmitButton,
+  SubmitError,
+} from '../../../components/form';
 import {
   FilmLookupItem,
   LiteracyMaterialLookupItem,
   SongLookupItem,
   StoryLookupItem,
 } from '../../../components/form/Lookup';
-import { renderAccordionSection } from './AccordionSection';
-import { ProductFormFragment } from './ProductForm.generated';
+import { AccordionSection } from './AccordionSection';
+import {
+  ProductFormFragment,
+  useDeleteProductMutation,
+} from './ProductForm.generated';
+
+const useStyles = makeStyles(({ spacing }) => ({
+  submissionBlurb: {
+    margin: spacing(4, 0),
+    width: spacing(50),
+    '& h4': {
+      marginBottom: spacing(1),
+    },
+  },
+  deleteButton: {
+    marginLeft: spacing(1),
+  },
+}));
 
 export interface ProductFormCustomValues {
   product: {
@@ -27,14 +53,63 @@ export interface ProductFormCustomValues {
 
 export const ProductForm = <FormMutationValues extends any>({
   product,
-  ...formProps
+  ...props
 }: FormProps<FormMutationValues> & {
   product?: ProductFormFragment;
 }) => {
+  const classes = useStyles();
+
+  const { enqueueSnackbar } = useSnackbar();
+  const navigate = useNavigate();
+  const [deleteProduct] = useDeleteProductMutation({
+    awaitRefetchQueries: true,
+    refetchQueries: [GQLOperations.Query.Engagement],
+  });
+  const handleDelete = async () => {
+    product &&
+      (await deleteProduct({
+        variables: {
+          productId: product.id,
+        },
+      }));
+
+    enqueueSnackbar(`Product deleted`, {
+      variant: 'success',
+    });
+
+    navigate('../../');
+  };
+
   return (
-    <Form<FormMutationValues>
-      {...formProps}
-      children={renderAccordionSection(product)}
-    ></Form>
+    <Form<FormMutationValues> {...props}>
+      {({ handleSubmit, ...rest }) => (
+        <form onSubmit={handleSubmit}>
+          <SubmitError />
+          <FieldGroup prefix="product">
+            <AccordionSection product={product} {...rest} />
+          </FieldGroup>
+          <div className={classes.submissionBlurb}>
+            <Typography variant="h4">Check Your Selections</Typography>
+            <Typography>
+              If the selections above look good to you, go ahead and save your
+              Product. If you need to edit your choices, do that above.
+            </Typography>
+          </div>
+          <SubmitButton fullWidth={false} color="primary" size="medium">
+            Save Product
+          </SubmitButton>
+          {product && (
+            <ErrorButton
+              fullWidth={false}
+              size="medium"
+              onClick={handleDelete}
+              className={classes.deleteButton}
+            >
+              Delete Product
+            </ErrorButton>
+          )}
+        </form>
+      )}
+    </Form>
   );
 };
